+++ src/test/java/us/bringardner/io/filesource/test/MemoryRandomAccessIoBufferTests.java
@@ -84,7 +84,7 @@ public class MemoryRandomAccessIoBufferTests extends FileSourceRandomAccessIoBuf
 		int doTries = 40;
 		try(IRandomAccessIoController buf = getRandomAccessFileStream(file)){
 			while(tries < doTries) {
-				long pos = r.nextLong(len);
+				long pos = r.nextInt((int)len);
 				int idx = ((int)pos) % testData.length;
 				int expect = testData[idx];
 				int i = buf.read(pos);
@@ -112,7 +112,7 @@ public class MemoryRandomAccessIoBufferTests extends FileSourceRandomAccessIoBuf
 		int doTries = 40;
 		try(IRandomAccessIoController buf = getRandomAccessFileStream(file)){
 			while(tries < doTries) {
-				long pos = r.nextLong(len);
+				long pos = r.nextInt((int)len);
 				int idx = ((int)pos) % data.length;
 				int expect = testData[idx];
 				buf.write(pos, (byte) expect);
